# base_app/static/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a REST API that consumes the Yelp and Google Map APIs. A restaurant recommendation with details is returned.",
    "version" : "1.0.0",
    "title" : "Hangry Ateball",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "hangry-ateball-api.herokuapp.com",
  "basePath" : "/api/v1",
  "schemes" : [ "http" ],
  "paths" : {
    "/random" : {
      "get" : {
        "summary" : "search for a random restaurant recommendation",
        "description" : "This will return a random restaurant recommendation.\n",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "latitude",
          "in" : "query",
          "description" : "If an address is not provided, latitude and longitude are required. Pass an integer representing the latitude of the location you would like to search from.",
          "required" : false,
          "type" : "number"
        }, {
          "name" : "longitude",
          "in" : "query",
          "description" : "If an address is not provided, latitude and longitude are required. Pass an integer representing the longitude of the location you would like to search from.",
          "required" : false,
          "type" : "number"
        }, {
          "name" : "address",
          "in" : "query",
          "description" : "If latitude and longitude are not provided, an address is required. A full address would include a street, city, state, and zip code (12345 Rodeo Dr, Los Angeles, CA 90210). Zip code is not required. An address can also include only a city and state (Los Angeles, CA) or state (CA).",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria"
          },
          "400" : {
            "description" : "invalid input"
          }
        }
      }
    },
    "/recommendations" : {
      "get" : {
        "summary" : "search for a restaurant recommendation based on preferences",
        "description" : "By passing in the appropriate options, you can search for\na restaurant recommendation.\n",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "latitude",
          "in" : "query",
          "description" : "If an address is not provided, latitude and longitude are required. Pass an integer representing the latitude of the location you would like to search from.",
          "required" : false,
          "type" : "number"
        }, {
          "name" : "longitude",
          "in" : "query",
          "description" : "If an address is not provided, latitude and longitude are required. Pass an integer representing the longitude of the location you would like to search from.",
          "required" : false,
          "type" : "number"
        }, {
          "name" : "address",
          "in" : "query",
          "description" : "If latitude and longitude are not provided, an address is required. A full address would include a street, city, state, and zip code (12345 Rodeo Dr, Los Angeles, CA 90210). Zip code is not required. An address can also include only a city and state (Los Angeles, CA) or state (CA).",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "price",
          "in" : "query",
          "description" : "Pass an integer (1 - 4) to select a maximum price point ($ - $$$$). For example, when 3 is entered, the search will also include restaurants with a price point of 2 and 1.",
          "required" : false,
          "type" : "integer",
          "minimum" : 1,
          "maximum" : 4
        }, {
          "name" : "travel",
          "in" : "query",
          "description" : "Enter either 'walk' or 'drive'. If walk is entered, a radius of 1600 meters will be searched. If drive is entered, a radius of 16,000 meters will be searched.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "categories",
          "in" : "query",
          "description" : "Pass a selection of cuisine types to filter restaurants with. The category filter can be a list of comma-delimited categories.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria"
          },
          "400" : {
            "description" : "invalid input"
          }
        }
      }
    }
  },
  "definitions" : { }
}